Add minor updates to finalise
