+++ data/tasks.txt
@@ -1,2 +1,1 @@
 T;;;1;;;hello
-T;;;1;;;bye bye
