--- src/main/java/duke/data/exception/DukeException.java
@@ -1,10 +1,0 @@
-package duke.data.exception;
-
-/**
- * Exception class for errors related to Duke
- */
-public class DukeException extends Exception {
-    public DukeException(String errorDesc) {
-        super(errorDesc);
-    }
-}
