+++ src/main/java/megatron/data/exception/DukeException.java
@@ -1,0 +1,10 @@
+package megatron.data.exception;
+
+/**
+ * Exception class for errors related to Duke
+ */
+public class DukeException extends Exception {
+    public DukeException(String errorDesc) {
+        super(errorDesc);
+    }
+}
