+++ src/main/resources/view/MainWindow.fxml
@@ -3,13 +3,10 @@
 <?import javafx.scene.control.Button?>
 <?import javafx.scene.control.ScrollPane?>
 <?import javafx.scene.control.TextField?>
-<?import javafx.scene.effect.Light.Distant?>
-<?import javafx.scene.effect.Lighting?>
-<?import javafx.scene.effect.Shadow?>
 <?import javafx.scene.layout.AnchorPane?>
 <?import javafx.scene.layout.VBox?>
 
-<AnchorPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="600.0" prefWidth="400.0" xmlns="http://javafx.com/javafx/21" xmlns:fx="http://javafx.com/fxml/1" fx:controller="duke.MainWindow">
+<AnchorPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="600.0" prefWidth="400.0" xmlns="http://javafx.com/javafx/21" xmlns:fx="http://javafx.com/fxml/1" fx:controller="megatron.MainWindow">
     <children>
         <TextField fx:id="userInput" layoutX="6.0" layoutY="562.0" onAction="#handleUserInput" prefHeight="33.0" prefWidth="312.0" AnchorPane.bottomAnchor="5.0" />
         <Button fx:id="sendButton" layoutX="324.0" layoutY="563.0" mnemonicParsing="false" onAction="#handleUserInput" prefHeight="31.0" prefWidth="71.0" text="Send" />
